move `delim_only?` method to `Transforms::Helpers`
