+++ lib/kiba/extend/transforms/clean.rb
@@ -6,33 +6,15 @@ module Kiba
       # Transformations to clean up data
       module Clean
         ::Clean = Kiba::Extend::Transforms::Clean
-
-        # Mixins for Clean transformations
-        module Helpers
-          extend self
-          # Indicates whether a field value is delimiter-only. If `usenull` is set to true, the
-          #   %NULLVALUE% string is treated as empty in detecting delimiter-only-ness
-          # @param val [String] The field value to check
-          # @param delim [String] The multivalue delimiter
-          # @param usenull [Boolean] If true, replaces '%NULLVALUE%' with '' to make determination
-          # @return [false] if `value` is nil, empty, or contains characters other than delimiter(s)
-          #   and leading/trailing spaces
-          # @return [true] if `value` contains only delimiter(s) and leading/trailing spaces
-          def delim_only?(val, delim, usenull = false)
-            return false if val.nil?
-            return false if val.strip.empty?
-            
-            chk = val.gsub(delim, '').strip
-            chk = chk.gsub('%NULLVALUE%', '').strip if usenull
-            chk.empty?
-          end
-        end
-
+        
+        # Sorts the multiple values within a field alphabetically
         class AlphabetizeFieldValues
-          include Clean::Helpers
-          def initialize(fields:, delim:)
+          include Kiba::Extend::Transforms::Helpers
+          def initialize(fields:, delim:, usenull: false, direction: :asc)
             @fields = fields
             @delim = delim
+            @usenull = usenull
+            @direction = direction
           end
 
           # @private
@@ -71,7 +53,7 @@ module Kiba
         end
 
         class DelimiterOnlyFields
-          include Clean::Helpers
+          include Kiba::Extend::Transforms::Helpers
           def initialize(delim:, use_nullvalue: false)
             @delim = delim
             @use_nullvalue = use_nullvalue
